+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 - Features:
 
   - Use NOS-Brick to setup OS-Brick for non OpenStack usage.
+  - Can setup persistence directly to use key-value storage.
 
 0.2.1 (2018-06-14)
 ------------------